Prints the constraints errors on CI
